Renames Demo from AwesomeTSProject to Raygun4ReactNativeDemo
