+++ demo/android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.awesometsproject",
+    package = "com.raygun4reactnativedemo",
 )
 
 android_resource(
     name = "res",
-    package = "com.awesometsproject",
+    package = "com.raygun4reactnativedemo",
     res = "src/main/res",
 )
