+++ demo/android/app/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.awesometsproject">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.raygun4reactnativedemo">
 	<uses-permission android:name="android.permission.INTERNET" />
 	<application android:name=".MainApplication" android:label="@string/app_name" android:icon="@mipmap/ic_launcher" android:roundIcon="@mipmap/ic_launcher_round" android:allowBackup="false" android:theme="@style/AppTheme">
 		<service android:name="com.raygun.raygun4android.services.CrashReportingPostService" android:exported="false" android:permission="android.permission.BIND_JOB_SERVICE" android:process=":crashreportingpostservice" />
