+++ demo/android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'AwesomeTSProject'
+rootProject.name = 'Raygun4ReactNativeDemo'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
 include ':app'
