+++ demo/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "AwesomeTSProject",
-  "displayName": "AwesomeTSProject"
+  "name": "Raygun4ReactNativeDemo",
+  "displayName": "Raygun4ReactNativeDemo"
 }
