+++ demo/ios/Raygun4ReactNativeDemo.xcodeproj/xcshareddata/xcschemes/AwesomeTSProject-tvOS.xcscheme
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "AwesomeTSProject.app"
-               BlueprintName = "AwesomeTSProject"
-               ReferencedContainer = "container:AwesomeTSProject.xcodeproj">
+               BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+               BuildableName = "Raygun4ReactNativeDemo-tvOS.app"
+               BlueprintName = "Raygun4ReactNativeDemo-tvOS"
+               ReferencedContainer = "container:Raygun4ReactNativeDemo.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -32,10 +32,10 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "AwesomeTSProjectTests.xctest"
-               BlueprintName = "AwesomeTSProjectTests"
-               ReferencedContainer = "container:AwesomeTSProject.xcodeproj">
+               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
+               BuildableName = "Raygun4ReactNativeDemo-tvOSTests.xctest"
+               BlueprintName = "Raygun4ReactNativeDemo-tvOSTests"
+               ReferencedContainer = "container:Raygun4ReactNativeDemo.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -54,10 +54,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "AwesomeTSProject.app"
-            BlueprintName = "AwesomeTSProject"
-            ReferencedContainer = "container:AwesomeTSProject.xcodeproj">
+            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+            BuildableName = "Raygun4ReactNativeDemo-tvOS.app"
+            BlueprintName = "Raygun4ReactNativeDemo-tvOS"
+            ReferencedContainer = "container:Raygun4ReactNativeDemo.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -71,10 +71,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "AwesomeTSProject.app"
-            BlueprintName = "AwesomeTSProject"
-            ReferencedContainer = "container:AwesomeTSProject.xcodeproj">
+            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+            BuildableName = "Raygun4ReactNativeDemo-tvOS.app"
+            BlueprintName = "Raygun4ReactNativeDemo-tvOS"
+            ReferencedContainer = "container:Raygun4ReactNativeDemo.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
