+++ demo/ios/Raygun4ReactNativeDemo.xcodeproj/xcshareddata/xcschemes/AwesomeTSProject.xcscheme
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "AwesomeTSProject-tvOS.app"
-               BlueprintName = "AwesomeTSProject-tvOS"
-               ReferencedContainer = "container:AwesomeTSProject.xcodeproj">
+               BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+               BuildableName = "Raygun4ReactNativeDemo.app"
+               BlueprintName = "Raygun4ReactNativeDemo"
+               ReferencedContainer = "container:Raygun4ReactNativeDemo.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -32,10 +32,10 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "AwesomeTSProject-tvOSTests.xctest"
-               BlueprintName = "AwesomeTSProject-tvOSTests"
-               ReferencedContainer = "container:AwesomeTSProject.xcodeproj">
+               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
+               BuildableName = "Raygun4ReactNativeDemoTests.xctest"
+               BlueprintName = "Raygun4ReactNativeDemoTests"
+               ReferencedContainer = "container:Raygun4ReactNativeDemo.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -54,10 +54,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "AwesomeTSProject-tvOS.app"
-            BlueprintName = "AwesomeTSProject-tvOS"
-            ReferencedContainer = "container:AwesomeTSProject.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "Raygun4ReactNativeDemo.app"
+            BlueprintName = "Raygun4ReactNativeDemo"
+            ReferencedContainer = "container:Raygun4ReactNativeDemo.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -71,10 +71,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "AwesomeTSProject-tvOS.app"
-            BlueprintName = "AwesomeTSProject-tvOS"
-            ReferencedContainer = "container:AwesomeTSProject.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "Raygun4ReactNativeDemo.app"
+            BlueprintName = "Raygun4ReactNativeDemo"
+            ReferencedContainer = "container:Raygun4ReactNativeDemo.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
