+++ demo/ios/Raygun4ReactNativeDemo.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:AwesomeTSProject.xcodeproj">
+      location = "group:Raygun4ReactNativeDemo.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
