+++ demo/ios/Raygun4ReactNativeDemo.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
