+++ demo/ios/Raygun4ReactNativeDemo/AppDelegate.m
@@ -38,7 +38,7 @@ static void InitializeFlipper(UIApplication *application) {
   RCTBridge *bridge = [[RCTBridge alloc] initWithDelegate:self
                                             launchOptions:launchOptions];
   RCTRootView *rootView = [[RCTRootView alloc] initWithBridge:bridge
-                                                   moduleName:@"AwesomeTSProject"
+                                                   moduleName:@"Raygun4ReactNativeDemo"
                                             initialProperties:nil];
 
   rootView.backgroundColor = [[UIColor alloc] initWithRed:1.0f
