+++ demo/ios/Raygun4ReactNativeDemo/Images.xcassets/AppIcon.appiconset/Contents.json
