+++ demo/ios/Raygun4ReactNativeDemo/Images.xcassets/Contents.json
