+++ demo/ios/Raygun4ReactNativeDemoTests/Raygun4ReactNativeDemo.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface AwesomeTSProjectTests : XCTestCase
+@interface Raygun4ReactNativeDemoTests : XCTestCase
 
 @end
 
-@implementation AwesomeTSProjectTests
+@implementation Raygun4ReactNativeDemoTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
